split atomic tests to a dedicated file
